% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{main}
\usage{
main(input_table_path, xmlurl)
}
\arguments{
\item{input_table_path}{input_table_path}

\item{xmlurl}{xmlurl}
}
\description{
main function in this package
}
\author{
Yiming Zhao
}
