% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mirror.R
\name{plot_mirror}
\alias{plot_mirror}
\title{plot_mirror}
\usage{
plot_mirror(fileName, f.msms, xmlurl, min_intensity = 100, cex = 1,
  srt = 0, ppm = 20, PPM_denominator = 1e+06, pdf_width = 20,
  pdf_height = 10)
}
\arguments{
\item{fileName}{filename}

\item{f.msms}{f.msms}

\item{xmlurl}{modifications.xml path}

\item{min_intensity}{min_intensity=100}

\item{cex}{cex}

\item{srt}{srt}

\item{ppm}{ppm}

\item{PPM_denominator}{PPM_denominator}

\item{pdf_width}{pdf_width}

\item{pdf_height}{pdf_height}
}
\description{
plot_mirror function in this package
}
\author{
Yiming Zhao
}
