% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_colors.R
\name{r.color}
\alias{r.color}
\title{r.color}
\usage{
r.color(i)
}
\description{
r.color
}
\seealso{
Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}
}

