% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_colors.R
\name{r.color.setAlpha}
\alias{r.color.setAlpha}
\title{r.color.setAlpha}
\usage{
r.color.setAlpha(color, alpha)
}
\description{
r.color.setAlpha
}
\seealso{
Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

\code{\link{r.color}} \code{\link{r.color.gradient}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}
}

