% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.divideplot}
\alias{r.divideplot}
\title{r.divideplot}
\usage{
r.divideplot(nrow = 1, ncol = 1)
}
\seealso{
Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}

