\name{r.export.dataanalysis}
\alias{r.export.dataanalysis}
\title{r.plot.bar}
\usage{
  r.export.dataanalysis(values = NULL, col = NULL,
    horizontal = FALSE, label.cex = 0.7,
    label.rotation = 45, label.adjX = 1.1,
    label.adjY = 1.1, legend = TRUE,
    legend.pos = "topright", legend.pch = 15,
    legend.cex = 0.6,
    legend.backgroundCol = rgb(0.98, 0.98, 1, 0.97),
    background = T, grid = T,
    backgroundCol = rgb(0.85, 0.85, 0.9),
    foregroundCol = rgb(0.95, 0.95, 1), ...)
}
\description{
  r.plot.bar
}

