% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_report.R
\name{r.export.dataoverview}
\alias{r.export.dataoverview}
\title{r.export.dataoverview}
\usage{
r.export.dataoverview(dades, col = NULL, horizontal = FALSE,
  label.cex = 0.7, label.rotation = 45, legend = FALSE,
  legend.pos = "topright", legend.cex = 0.6, folder = "output/",
  filePrefix = "img_", filePrefixCont = "cont_", filePrefixCat = "cat_",
  fileSufix = "", width = 480, height = 480, ...)
}
\description{
r.export.dataoverview
}

