% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.ggplot.bar}
\alias{r.ggplot.bar}
\title{r.ggplot.bar}
\usage{
r.ggplot.bar(df, x, y, angle = 65, order = FALSE, orderType = "desc",
  color = rgb(0.3, 0.3, 0.6, 0.8), stroke = rgb(0, 0, 0, 0.2),
  title = NULL, xlab = NULL, ylab = NULL, stat = "identity")
}
\description{
r.ggplot.bar
}

