% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rggplot_corr.R
\name{r.ggplot.corr}
\alias{r.ggplot.corr}
\title{r.ggplot.corr}
\usage{
r.ggplot.corr(Mcorr, listVar = colnames(Mcorr), color = rgb(0.3, 0.3, 0.6,
  0.8), stroke = rgb(0, 0, 0, 0.2), title = NULL, xlab = NULL,
  ylab = NULL, showText = NULL, angle = 60)
}
\arguments{
\item{Mcorr}{Matrix of correlations (cor).}
}
\description{
Add data to the current plot.
}
\seealso{
\code{\link{cor}}
}

