% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.ggplot.density}
\alias{r.ggplot.density}
\title{r.ggplot.density}
\usage{
r.ggplot.density(df, var, color = r.color(1), stroke = rgb(0, 0, 0, 0.2),
  title = NULL, xlab = NULL, ylab = NULL)
}
\description{
r.ggplot.density
}

