% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rggplot.R
\name{r.ggplot}
\alias{r.ggplot}
\title{r.ggplot}
\usage{
r.ggplot(df, x, y, logScale = FALSE, color = rgb(0.3, 0.3, 0.6, 0.2),
  title = NULL, xlab = NULL, ylab = NULL, size = 1.5)
}
\description{
r.ggplot
}

