% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.gplot.boxplot}
\alias{r.gplot.boxplot}
\title{r.gplot.boxplot}
\usage{
r.gplot.boxplot(y, x = factor(0), angle = 65, order = FALSE,
  orderRev = FALSE, color = rplot:::param.color.bar, stroke = rgb(0, 0, 0,
  0.2), title = NULL, xlab = NULL, ylab = NULL, stat = "boxplot")
}
\description{
r.gplot.boxplot
}

