% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggplot.R
\name{r.gplot}
\alias{r.gplot}
\title{r.gplot}
\usage{
r.gplot(x, y, logScale = FALSE, logScaleX = logScale,
  logScaleY = logScale, color = rgb(0.3, 0.3, 0.6, 0.2), title = NULL,
  xlab = NULL, ylab = NULL, size = 1.5, type = "p")
}
