% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_interactive.R
\name{r.int.kmeans}
\alias{r.int.kmeans}
\title{r.int.kmeans}
\usage{
r.int.kmeans(x, fmin = 0.05, fmax = 1, fstep = 0.05, main = "",
  sub = "", xlab = "", ylab = "")
}
\description{
r.int.kmeans
}

