% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_interactive.R
\name{r.int.plot.smoothkmeans}
\alias{r.int.plot.smoothkmeans}
\title{r.int.plot.smoothkmeans}
\usage{
r.int.plot.smoothkmeans(x, main = "", sub = "", xlab = "", ylab = "")
}
\description{
r.int.plot.smoothkmeans
}

