% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_interactive.R
\name{r.int.plot2D.pca}
\alias{r.int.plot2D.pca}
\title{r.int.plot2D.pca}
\usage{
r.int.plot2D.pca(x = NULL, pca = NULL, clustReal = NULL,
  clustModel = NULL, main = "", sub = "", xlab = "", ylab = "",
  xaxis = T, yaxis = T, box = T, step = 1, ...)
}
\description{
r.int.plot2D.pca
}

