% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_interactive.R
\name{r.iplot.kmeans.shapes}
\alias{r.iplot.kmeans.shapes}
\title{r.iplot.kmeans.shapes}
\usage{
r.iplot.kmeans.shapes(x, fmin = 0.05, fmax = 1, fstep = 0.05,
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL)
}
