% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.palette}
\alias{r.palette}
\title{r.palette}
\usage{
r.palette()
}
\description{
r.palette
}
\seealso{
\code{\link{r.plot.new}} \code{\link{r.plot}} \code{\link{r.plot.add}}
\cr 2nd axis: \code{\link{r.plot.setmargins}} \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
\cr Colors: \code{\link{r.palette}} \code{\link{r.color}}
\cr Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

