% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.palette.show}
\alias{r.palette.show}
\title{r.palette.show}
\usage{
r.palette.show(palette = NULL, alpha = NULL)
}
\description{
r.palette.show
}
\seealso{
\code{\link{r.plot}} \code{\link{r.plot.add}}
\cr 3rd axis: \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
\cr Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}} \code{\link{r.color}} \code{\link{r.color.gradient}}
\cr Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

