% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.performance.metrics}
\alias{r.performance.metrics}
\title{r.performance.metrics}
\usage{
r.performance.metrics(score, target, threshold = 0.5, round = 5,
  big.mark = ".", decimal.mark = ",", scientific = FALSE,
  printConfMat = TRUE, printF1 = TRUE)
}
\description{
r.performance.metrics
}

