% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.add.expression}
\alias{r.plot.add.expression}
\title{r.plot.add.expression}
\usage{
r.plot.add.expression(expression)
}
\arguments{
\item{expression}{Expression to execute with last plot device.}
}
\description{
Add data to the current plot.
}
\seealso{
Main: \code{\link{r.plot}} \code{\link{r.plot.add}} \code{\link{r.plot.new}}

3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}

Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}

