\name{r.plot.add}
\alias{r.plot.add}
\title{r.plot.add}
\usage{
  r.plot.add(y, x = NULL, type = "l", lwd = 1, pch = 16,
    cex = 1, icol = 1, col = NULL, ...)
}
\description{
  r.plot.add
}
\seealso{
  \code{\link{r.plot.new}} \code{\link{r.plot}}
  \code{\link{r.plot.add}} \cr 2nd axis:
  \code{\link{r.plot.setmargins}}
  \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
  \cr Colors: \code{\link{r.palette}} \code{\link{r.color}}
  \cr Tools: \code{\link{r.plot.window}}
  \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

