\name{r.plot.addaxis}
\alias{r.plot.addaxis}
\title{r.plot.addaxis}
\usage{
  r.plot.addaxis(values = NULL, range = NULL, at = NULL,
    text = NULL, side = 4, line = 3, ...)
}
\description{
  r.plot.addaxis
}
\seealso{
  \code{\link{r.plot.new}} \code{\link{r.plot}}
  \code{\link{r.plot.add}} \cr 2nd axis:
  \code{\link{r.plot.setmargins}}
  \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
  \cr Colors: \code{\link{r.palette}} \code{\link{r.color}}
  \cr Tools: \code{\link{r.plot.window}}
  \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

