% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_dr_vs_ar.R
\name{r.plot.ar}
\alias{r.plot.ar}
\title{r.plot.ar}
\usage{
r.plot.ar(score, target, weight = rep(1, length(target)), npoints = 200,
  main = "Acceptance", sub = NULL, xlab = "cutoff",
  ylab = "\% Acceptance", icol = 1, col = r.color(icol),
  target_value = 1, ...)
}
