\name{r.plot.burbujas}
\alias{r.plot.burbujas}
\title{r.plot.burbujas}
\usage{
  r.plot.burbujas(datos, segmentacion, target,
    relativeToMean = FALSE, relativeToMeanX = FALSE,
    relativeToMeanY = FALSE, logScale = FALSE,
    logScaleX = FALSE, logScaleY = FALSE,
    laplaceSmooth = FALSE, showMean = FALSE,
    showMeanX = FALSE, showMeanY = FALSE,
    showTargetAbsMean = FALSE, sizeMin = 1, sizeMax = 3,
    xlab = "Vol.", ylab = "Freq.Rel.Target", ...)
}
\description{
  r.plot.burbujas
}

