% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.confusionmatrix}
\alias{r.plot.confusionmatrix}
\title{r.plot.confusionmatrix}
\usage{
r.plot.confusionmatrix(score, target, threshold = 0.5, round = 5,
  big.mark = ".", decimal.mark = ",", scientific = FALSE)
}

