% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.coord.axis}
\alias{r.plot.coord.axis}
\title{r.plot.coord.axis}
\usage{
r.plot.coord.axis(values = NULL, range = NULL, at = NULL, text = NULL,
  side = 4, line = 3, axisCol = param.color.axis, ...)
}
\arguments{
\item{values}{values for the new axis scale, it is supposed to be normal data like x or y, but it will be used for a new axis z.}
}
\description{
Paints the new axis. Useful to create a 3rd axis.
}
\details{
﻿
}
\seealso{
Main: \code{\link{r.plot}} \code{\link{r.plot.add}} \code{\link{r.plot.new}}

3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}
}

