% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_dr_vs_ar.R
\name{r.plot.dr}
\alias{r.plot.dr}
\title{r.plot.dr}
\usage{
r.plot.dr(score, target, weight = rep(1, length(target)), npoints = 200,
  main = "Default", sub = NULL, xlab = "cutoff", ylab = "\% Default",
  icol = 1, col = r.color(icol), target_value = 1, ...)
}
