% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_dr_vs_ar.R
\name{r.plot.dr_vs_ar}
\alias{r.plot.dr_vs_ar}
\title{r.plot.dr_vs_ar}
\usage{
r.plot.dr_vs_ar(score, target, weight = rep(1, length(target)),
  npoints = 200, main = "DR vs AR", sub = NULL, xlab = "\% Acceptance",
  ylab = "\% Default", icol = 1, icol.max = 11, col = r.color(icol),
  col.max = r.color(icol.max), showMax = TRUE, showRandom = TRUE,
  xfactor = 1, target_value = 1, ...)
}
