% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.F1}
\alias{r.plot.F1}
\title{r.plot.F1}
\usage{
r.plot.F1(score, target, threshold = 0.5, round = 5, big.mark = ".",
  decimal.mark = ",", scientific = FALSE)
}
\description{
r.plot.F1
}

