% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.gain.old}
\alias{r.plot.gain.old}
\title{r.plot.gain.old}
\usage{
r.plot.gain.old(score, fuga, npoints = 20, perc = 0.2, mode = 0,
  sub = "", icol = 1, col = NULL, ...)
}
\arguments{
\item{mode}{1 Optimista, -1 Pesimista, 0 Sense Modificar, 2 Area, 3 Random}
}
\description{
r.plot.gain.old
}

