% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_gain.R
\name{r.plot.gain}
\alias{r.plot.gain}
\title{r.plot.gain}
\usage{
r.plot.gain(score, target, npoints = 20, mode = "avg",
  main = "Gain Curve", sub = NULL, icol = 1, icol.max = 11,
  col = r.color(icol), col.max = r.color(icol.max), showMessage = TRUE,
  showMax = TRUE, target_value = NULL, ...)
}
\arguments{
\item{mode}{\cr "def" = As it is
\cr "rnd" = Random
\cr "pos" = Optimist/Positive (Maximum)
\cr "neg" = Pessimist/Negative (Minimum)
\cr "avg" = Average of Optimist and Pessimist
\cr "area" = Area between Optimist and Pessimist}
}

