% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_graph.R
\name{r.plot.graph.text}
\alias{r.plot.graph.text}
\title{r.plot.graph.text}
\usage{
r.plot.graph.text(mat, textLabels = NULL, quantileCutOff = 0.6,
  quantileShowOff = 0.7, edgeWidth = 1, edgeWidthMin = edgeWidth,
  edgeWidthMax = edgeWidth, edgeColorAlpha = 0.5,
  edgeColorAlphaMin = edgeColorAlpha, edgeColorAlphaMax = edgeColorAlpha,
  vertexLabelColorDefault = rgb(0.3, 0.3, 0.3, 0.5),
  vertexLabelColor = NULL, palette = RColorBrewer::brewer.pal(name =
  "Dark2", n = 8), vertexLabelFont = 2, vertexLabelCex = 0.4,
  vertexShape = "none", vertexSize = 0, ...)
}
