% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_heatmap.R
\name{r.plot.heatmap}
\alias{r.plot.heatmap}
\title{r.plot.heatmap}
\usage{
r.plot.heatmap(x, y, z, levels = 10, palette = NULL, alpha = 0.4,
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL, xaxis = TRUE,
  yaxis = TRUE, box = TRUE, ...)
}
\description{
r.plot.heatmap
}

