% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_data.R
\name{r.plot.histogram}
\alias{r.plot.histogram}
\title{r.plot.histogram}
\usage{
r.plot.histogram(values = NULL, h = NULL, breaks = "Sturges",
  include.lowest = TRUE, right = TRUE, icol = 1, col = NULL, ...)
}
\description{
r.plot.histogram
}
\seealso{
\code{\link{hist}} \cr Others: \cr
\code{\link{r.plot.bar}} \code{\link{r.plot.histogram}} \code{\link{r.plot.distribution}} \code{\link{r.plot.data}} \cr
\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.target.table}} \cr
\code{\link{barplot}} \code{\link{hist}} \code{\link{bkde}}
}

