% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_kmeans.R
\name{r.plot.kmeans.shapes}
\alias{r.plot.kmeans.shapes}
\title{r.plot.kmeans.shapes}
\usage{
r.plot.kmeans.shapes(x, km = NULL, nclusters = 2, paintCentroids = F,
  filtrat = 1, nMax = NULL, main = NULL, sub = NULL, xlab = NULL,
  ylab = NULL, type = "l", lwd = 1, pch = 16, cex = 1)
}
\description{
r.plot.kmeans.shapes
}

