% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.legend}
\alias{r.plot.legend}
\title{r.plot.legend}
\usage{
r.plot.legend(labels = NULL, legend.pos = "topright", legend.col = NULL,
  legend.text.col = NULL, legend.ncol = 2, legend.pch = 16,
  legend.cex = 0.6, legend.lwd = 1,
  legend.backgroundCol = rplot:::param.color.legend.background,
  legend.rev = FALSE, legend.bty = "o", boxCol = param.color.box, ...)
}
\arguments{
\item{x}{Array of data for x axis.}

\item{y}{Array of data for y axis.}
}
\description{
Adds a legend.
}
\seealso{
3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}

Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}
