% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.lift}
\alias{r.plot.lift}
\title{r.plot.lift}
\usage{
r.plot.lift(score, target, npoints = 20, perc = 0.2, icol = 1,
  col = NULL, main = "Lift Curve", sub = NULL, ...)
}
\description{
r.plot.lift
}

