% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_lift.R
\name{r.plot.lifts}
\alias{r.plot.lifts}
\title{r.plot.lifts}
\usage{
r.plot.lifts(score.train, target.train, score.test, target.test, npoints = 20,
  icol = 1, col = r.color(icol), main = "Lift Curve", sub = NULL,
  showMessage = TRUE, target_value = NULL, ...)
}

