% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_graph.R
\name{r.plot.matrix.communities}
\alias{r.plot.matrix.communities}
\title{r.plot.matrix.communities}
\usage{
r.plot.matrix.communities(inputMatrix, quantileCutOff = 0.6, zlim = NULL,
  pal = NULL, main = "Matrix", xlab = NULL, ylab = "", ...)
}
\description{
r.plot.matrix.communities
}

