% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_graph.R
\name{r.plot.matrix}
\alias{r.plot.matrix}
\title{r.plot.matrix}
\usage{
r.plot.matrix(mat, zlim = c(min(mat), max(mat)), palette = NULL,
  centerZero = TRUE, main = NULL, xlab = NULL, ylab = NULL,
  sub = NULL, ...)
}
\description{
r.plot.matrix
}

