% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_radial.R
\name{r.plot.radial}
\alias{r.plot.radial}
\title{r.plot.radial}
\usage{
r.plot.radial(data, axis.labels = colnames(data)[-1], grid.min = NULL,
  grid.mid = NULL, grid.max = NULL, center.y = NULL,
  plot.extent.x = 1.2, plot.extent.y = 1.2, x.center.range = NULL,
  label.center.y = FALSE, grid.label.size = 3,
  gridline.label.offset = NULL, rounding = NULL,
  label.gridline.min = TRUE, label.gridline.mid = NULL,
  label.gridline.max = TRUE, label.color = rgb(127/255, 127/255, 127/255),
  axis.label.offset = 1.15, axis.label.size = 3, group.line.width = 0.1,
  group.point.size = 1, background.circle.color = rgb(229/255, 229/255,
  229/255), legend = if (nrow(data) > 1) TRUE else FALSE,
  legend.title = NULL, legend.text.size = grid.label.size)
}
\description{
Radial plot done by Paul Williamson (http://pcwww.liv.ac.uk/~william/)
}
\details{
r.plot.radial
}

