% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.restorepar}
\alias{r.plot.restorepar}
\title{r.plot.restorepar}
\usage{
r.plot.restorepar()
}
\seealso{
Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}
