% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_roc.R
\name{r.plot.roc}
\alias{r.plot.roc}
\title{r.plot.roc}
\usage{
r.plot.roc(score, target, main = "ROC Curve",
  xlab = "1-specificity    |    FPR    |    fp/(tn+fp)",
  ylab = "sensitivity    |    TPR    |    tp/(tp+fn)", sub = NULL,
  fill = FALSE, colorFill = r.color(1), colorArea = r.color(1),
  showMessage = TRUE, ...)
}
