% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.rocs}
\alias{r.plot.rocs}
\title{r.plot.rocs}
\usage{
r.plot.rocs(score.train, target.train, score.test, target.test,
  main = "ROC Curve", xlab = "1-specificity    tn/(tn+fp)",
  ylab = "sensitivity    tp/(tp+fn)", sub = NULL, icol.train = 1,
  icol.test = 2, col.train = r.color(icol.train),
  col.test = r.color(icol.test), showMessage = TRUE, ...)
}

