% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_squash.R
\name{r.plot.squash}
\alias{r.plot.squash}
\title{r.plot.squash}
\usage{
r.plot.squash(x, y, z, logScale = FALSE, logScoreScale = FALSE,
  title = NULL, xlab = NULL, ylab = NULL, zlab = NULL, nx = 300,
  nz = 30, FUN = mean, colFn = jet, shrink = 10, xlim = NULL,
  ylim = NULL)
}
