% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_wgain.R
\name{r.plot.wgains}
\alias{r.plot.wgains}
\title{r.plot.wgains}
\usage{
r.plot.wgains(score.train, target.train, weight.train, score.test, target.test,
  weight.test, target_value = 1, target_w.train = ifelse(target.train ==
  target_value, weight.train, 0), target_w.test = ifelse(target.test ==
  target_value, weight.test, 0), npoints = 20, mode = "avg",
  main = "Gain Curve", sub = NULL, icol.train = 1, icol.test = 2,
  icol.max = 11, col.train = r.color(icol.train),
  col.test = r.color(icol.test), col.max = r.color(icol.max),
  showMessage = TRUE, showMax = TRUE, ...)
}
\arguments{
\item{mode}{\cr "def" = As it is
\cr "rnd" = Random
\cr "pos" = Optimist/Positive (Maximum)
\cr "neg" = Pessimist/Negative (Minimum)
\cr "avg" = Average of Optimist and Pessimist}
}

