% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.window}
\alias{r.plot.window}
\title{r.plot.window}
\usage{
r.plot.window(width = 900, height = 900, use.win.graph = FALSE, ...)
}
\description{
r.plot.window
}
\seealso{
Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}

