% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_dimred.R
\name{r.plot2D.data}
\alias{r.plot2D.data}
\title{r.plot2D.data}
\usage{
r.plot2D.data(x, clustReal = NULL, clustModel = NULL, comp1 = 1,
  comp2 = 2, xlim = c(0, 1), ylim = c(0, 1), main = NULL, sub = NULL,
  xlab = NULL, ylab = NULL, xaxis = T, yaxis = T, box = T)
}
\description{
r.plot2D.data
}

