% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_clusters.R
\name{r.plot2D.nn}
\alias{r.plot2D.nn}
\title{r.plot2D.nn}
\usage{
r.plot2D.nn(x, clustReal = NULL, clustModel = NULL, xlim = c(0, 1),
  ylim = c(0, 1), main = "", sub = "", xlab = "", ylab = "",
  xaxis = T, yaxis = T, box = T)
}
\description{
r.plot2D.nn
}

