% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.setAlpha}
\alias{r.setAlpha}
\title{r.setAlpha}
\usage{
r.setAlpha(color, alpha)
}
\description{
r.setAlpha
}
\seealso{
\code{\link{r.plot}} \code{\link{r.plot.add}}
\cr Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}} \code{\link{r.color}} \code{\link{r.color.gradient}}
}

