% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/params.R
\name{r.setPalette}
\alias{r.setPalette}
\title{r.setPalette}
\usage{
r.setPalette(palette)
}
\description{
r.setPalette
}
\seealso{
\code{\link{r.palette.set}}

Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

\code{\link{r.color}} \code{\link{r.color.gradient}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}
}

