\name{r.treemap}
\alias{r.treemap}
\title{r.treemap}
\usage{
  r.treemap(segment, segmentgroup, area, color,
    scale = NULL,
    lab = c(segmentgroup = TRUE, segment = FALSE),
    main = "Treemap", colorScaleLeft = rgb(0.9, 0, 0),
    colorScaleCenter = rgb(1, 1, 1),
    colorScaleRight = rgb(0, 0.9, 0),
    colorLabelsegmentgroup = rgb(0, 0, 0, 0.98),
    colorLabelSegment = rgb(0, 0, 0, 0.95),
    colorBorder = "black", cexMain = 1.5,
    cexLabelsegmentgroup = 1, cexLabelsegment = 0.1)
}
\description{
  r.treemap
}
\seealso{
  \code{\link{r.plot.new}} \code{\link{r.plot}}
  \code{\link{r.plot.add}} \cr 2nd axis:
  \code{\link{r.plot.setmargins}}
  \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
  \cr Other: \code{\link{r.treemap}}
  \code{\link{r.plot.bar}} \code{\link{r.plot.histogram}}
  \code{\link{r.plot.distribution}} \cr Colors:
  \code{\link{r.palette}} \code{\link{r.color}} \cr Tools:
  \code{\link{r.plot.window}} \code{\link{r.plot.reset}}
  \code{\link{r.plot.close}}
}

