% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_colors.R
\name{r.color.gradient}
\alias{r.color.gradient}
\title{r.color.gradient}
\usage{
r.color.gradient(z, levels = 10, palette = NULL)
}
\description{
r.color.gradient
}
\seealso{
Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}
}

