% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_colors.R
\name{r.color.setAlpha}
\alias{r.color.setAlpha}
\title{r.color.setAlpha}
\usage{
r.color.setAlpha(color, alpha)
}
\description{
r.color.setAlpha
}
\seealso{
Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}
}

