% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rggplot_corr.R
\name{r.ggplot.corr}
\alias{r.ggplot.corr}
\title{r.ggplot.corr}
\usage{
r.ggplot.corr(Mcorr, listVar = colnames(Mcorr), colorPos = rgb(70/255,
  130/255, 180/255), colorNeg = rgb(165/255, 42/255, 42/255),
  stroke = rgb(0, 0, 0, 0.2), title = NULL, xlab = NULL, ylab = NULL,
  showText = NULL, rowNames = NULL, colNames = NULL, angle = 60,
  expand = TRUE)
}
\arguments{
\item{Mcorr}{Matrix of correlations (cor).}
}
\description{
Add data to the current plot.
}
\seealso{
\code{\link{cor}}
}

