% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.ggplot.distribution}
\alias{r.ggplot.distribution}
\title{r.ggplot.distribution}
\usage{
r.ggplot.distribution(df, var, color = rgb(0.3, 0.3, 0.6, 0.8),
  stroke = rgb(0, 0, 0, 0.2), title = NULL, xlab = NULL, ylab = NULL)
}
\description{
r.ggplot.distribution
}

