% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.ggplot.histogram}
\alias{r.ggplot.histogram}
\title{r.ggplot.histogram}
\usage{
r.ggplot.histogram(df, var, color = rgb(0.3, 0.3, 0.6, 0.8), stroke = rgb(0,
  0, 0, 0.2), title = NULL, xlab = NULL, ylab = NULL)
}
\description{
r.ggplot.histogram
}

