% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.gplot.bar}
\alias{r.gplot.bar}
\title{r.gplot.bar}
\usage{
r.gplot.bar(x, y, angle = 65, order = FALSE, orderRev = FALSE,
  color = rplot:::param.color.bar, stroke = rgb(0, 0, 0, 0.2),
  title = NULL, xlab = NULL, ylab = NULL, stat = "identity")
}
