% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.gplot.distribution}
\alias{r.gplot.distribution}
\title{r.gplot.distribution}
\usage{
r.gplot.distribution(var, color = r.color(1), stroke = rgb(0, 0, 0, 0.2),
  title = NULL, xlab = NULL, ylab = NULL)
}
