% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rggplot_wrappers.R
\name{r.gplot.histogram}
\alias{r.gplot.histogram}
\title{r.gplot.histogram}
\usage{
r.gplot.histogram(var, color = r.color(1), stroke = rgb(0, 0, 0, 0.2),
  title = NULL, xlab = NULL, ylab = NULL)
}
\description{
r.gplot.histogram
}

