% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_interactive.R
\name{r.iplot2D.data}
\alias{r.iplot2D.data}
\title{r.iplot2D.data}
\usage{
r.iplot2D.data(x, clustReal = NULL, clustModel = NULL, main = NULL,
  sub = NULL, xlab = NULL, ylab = NULL, xaxis = T, yaxis = T,
  box = T, ...)
}
\description{
r.iplot2D.data
}

