% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_interactive.R
\name{r.iplot2D.pca}
\alias{r.iplot2D.pca}
\title{r.iplot2D.pca}
\usage{
r.iplot2D.pca(x = NULL, pca = NULL, clustReal = NULL, clustModel = NULL,
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL, xaxis = T,
  yaxis = T, box = T, ...)
}
