% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_colors.R
\name{r.palette.set}
\alias{r.palette.set}
\title{r.palette.set}
\usage{
r.palette.set(palette)
}
\description{
r.palette.set
}
\seealso{
\code{\link{r.setPalette}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}
}

