% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_colors.R
\name{r.palette.show}
\alias{r.palette.show}
\title{r.palette.show}
\usage{
r.palette.show(palette = NULL, alpha = NULL)
}
\description{
r.palette.show
}
\seealso{
Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

\code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}
}

