% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.add}
\alias{r.plot.add}
\title{r.plot.add}
\usage{
r.plot.add(x = NULL, y = NULL, type = "p", lwd = 1, pch = 16,
  cex = 1, icol = NULL, col = NULL, alpha = NULL, ...)
}
\description{
r.plot.add
}
\seealso{
\code{\link{r.plot.new}} \code{\link{r.plot}} \code{\link{r.plot.add}}
\cr 2nd axis: \code{\link{r.plot.setmargins}} \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
\cr Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}} \code{\link{r.color}} \code{\link{r.color.gradient}}
\cr Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

