% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.add}
\alias{r.plot.add}
\title{r.plot.add}
\usage{
r.plot.add(x = NULL, y = NULL, type = "p", lwd = 1, pch = 16,
  cex = 1, icol = NULL, col = NULL, alpha = NULL, ...)
}
\arguments{
\item{x}{Array of data for x axis.}

\item{y}{Array of data for y axis.}
}
\description{
Add data to the current plot.
}
\seealso{
3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}

Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}

