% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.add}
\alias{r.plot.add}
\title{r.plot.add}
\usage{
r.plot.add(x = NULL, y = NULL, type = "p", lwd = 1, pch = 16,
  cex = 1, icol = NULL, col = NULL, alpha = NULL, ...)
}
\arguments{
\item{x}{Array of data for x axis.}

\item{y}{Array of data for y axis.}
}
\description{
Add data to the current plot.
}
\details{
﻿
}
\seealso{
Main: \code{\link{r.plot}} \code{\link{r.plot.add}} \code{\link{r.plot.new}}

3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}

Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

\code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}

Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

