% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_data.R
\name{r.plot.bar}
\alias{r.plot.bar}
\title{r.plot.bar}
\usage{
r.plot.bar(values = NULL, table = NULL, labelsX = NULL, labelsY = NULL,
  labelsXDefault = FALSE, icol = NULL, col = NULL, horizontal = FALSE,
  beside = FALSE, label.cex = 0.7, label.rotation = 45,
  label.adjX = 1.1, label.adjY = 1.1, legend = TRUE,
  legend.pos = "topright", legend.pch = 15, legend.cex = 0.6,
  legend.backgroundCol = rgb(0.98, 0.98, 1, 0.97), background = T,
  grid = T, backgroundCol = rgb(229/255, 229/255, 229/255),
  foregroundCol = rgb(0.95, 0.95, 0.95), ...)
}
\description{
r.plot.bar
}
\seealso{
\code{\link{barplot}} \cr Others: \cr
\code{\link{r.plot.bar}} \code{\link{r.plot.histogram}} \code{\link{r.plot.distribution}} \code{\link{r.plot.data}} \cr
\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.target.table}} \cr
\code{\link{barplot}} \code{\link{hist}} \code{\link{bkde}}
}

