% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_bar.R
\name{r.plot.bar}
\alias{r.plot.bar}
\title{r.plot.bar}
\usage{
r.plot.bar(data, labelsX = NULL, labelsY = NULL, labelsXDefault = FALSE,
  icol = NULL, col = NULL, colBar = param.color.bar, horizontal = FALSE,
  beside = FALSE, thirdAxis = FALSE, main = NULL, sub = NULL,
  xlab = NULL, ylab = NULL, ylim = NULL, extraMarge = 0.02,
  label.cex = 0.7, label.rotation = 45, label.adjX = 1.1,
  label.adjY = 1.1, legend = NULL, legend.pos = "topright",
  legend.pch = 15, legend.cex = 0.6,
  legend.backgroundCol = param.color.legend.background, background = T,
  grid = T, backgroundCol = param.color.background,
  foregroundCol = param.color.foreground, axisCol = param.color.axis,
  axis = T, box = T, boxCol = param.color.box, useVector = FALSE,
  autoTranspose = TRUE, ...)
}
\description{
r.plot.bar
}
\seealso{
\code{\link{barplot}}

Wrappers: \code{\link{r.plot.bar}} \code{\link{r.plot.histogram}} \code{\link{r.plot.distribution}}

\code{\link{r.plot.heatmap}}

\code{\link{r.plot.matrix}} \code{\link{r.plot.matrix.communities}} \code{\link{r.plot.graph.text}}

\code{\link{r.plot.treemap}} \code{\link{r.plot.radial}}

\code{\link{r.plot.data}}

\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.target.table}}
}

