% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.coord}
\alias{r.plot.coord}
\title{r.plot.coord}
\usage{
r.plot.coord(x = NULL, y = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{x}{x axis scale.}

\item{y}{y axis scale.}
}
\description{
Defines a complete new coord system to the current plot.
}
\seealso{
3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}
}
