\name{r.plot.distribution}
\alias{r.plot.distribution}
\title{r.plot.distribution}
\usage{
  r.plot.distribution(x, bandwidth = 0.25,
    kernel = "epanech", range.x = NULL, truncate = TRUE,
    icol = 1, col = NULL, fill = TRUE, ...)
}
\description{
  r.plot.distribution
}
\seealso{
  \code{\link{bkde}} \cr Others: \cr
  \code{\link{r.plot.bar}} \code{\link{r.plot.histogram}}
  \code{\link{r.plot.distribution}}
  \code{\link{r.plot.data}} \cr \code{\link{r.binning}}
  \code{\link{r.segment}} \code{\link{r.segment.target}}
  \code{\link{r.segment.target.table}} \cr
  \code{\link{barplot}} \code{\link{hist}}
  \code{\link{bkde}}
}

