% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_distribution.R
\name{r.plot.distribution}
\alias{r.plot.distribution}
\title{r.plot.distribution}
\usage{
r.plot.distribution(x, bandwidth = 0.25, kernel = "epanech",
  range.x = NULL, truncate = TRUE, icol = 1, col = NULL, alpha = 0.5,
  colBorder = rgb(0, 0, 0), alphaBorder = 1, fill = TRUE, ...)
}
\seealso{
\code{\link{bkde}} \code{\link{density}} \code{\link{hist}}

Wrappers: \code{\link{r.plot.bar}} \code{\link{r.plot.histogram}} \code{\link{r.plot.distribution}}

\code{\link{r.plot.heatmap}}

\code{\link{r.plot.matrix}} \code{\link{r.plot.matrix.communities}} \code{\link{r.plot.graph.text}}

\code{\link{r.plot.treemap}} \code{\link{r.plot.radial}}

\code{\link{r.plot.data}}

\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.target.table}}
}
