% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model_risk.R
\name{r.plot.dr}
\alias{r.plot.dr}
\title{r.plot.dr}
\usage{
r.plot.dr(score, target, npoints = 20, icol = 1, col = r.color(icol),
  main = "Lift Curve", sub = NULL, showMessage = TRUE,
  target_value = NULL, ...)
}

