% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.gain}
\alias{r.plot.gain}
\title{r.plot.gain}
\usage{
r.plot.gain(score, target, npoints = 20, perc = 0.2, mode = "def",
  main = "Gain Curve", sub = NULL, icol = 1, col = NULL, ...)
}
\arguments{
\item{mode}{\cr "def" = As it is
\cr "rnd" = Random
\cr "pos" = Optimist/Positive (Maximum)
\cr "neg" = Pessimist/Negative (Minimum)
\cr "avg" = Average of Optimist and Pessimist
\cr "area" = Area between Optimist and Pessimist}
}
\description{
r.plot.gain
}

