% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.gains}
\alias{r.plot.gains}
\title{r.plot.gains}
\usage{
r.plot.gains(score.train, target.train, score.test, target.test, npoints = 20,
  mode = "def", main = "Gain Curve", sub = NULL, icol.train = 1,
  icol.test = 2, col.train = r.color(icol.train),
  col.test = r.color(icol.test), showMessage = TRUE, target_value = NULL,
  ...)
}
\arguments{
\item{mode}{\cr "def" = As it is
\cr "rnd" = Random
\cr "pos" = Optimist/Positive (Maximum)
\cr "neg" = Pessimist/Negative (Minimum)
\cr "avg" = Average of Optimist and Pessimist}
}

