% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_heatmap.R
\name{r.plot.heatmap}
\alias{r.plot.heatmap}
\title{r.plot.heatmap}
\usage{
r.plot.heatmap(x = NULL, y = NULL, z = NULL, agregateFunc = NULL,
  matrixData = NULL, xbreaks = 20, ybreaks = 20, levels = 12,
  smooth = 0, palette = param.palette.heatmap, main = NULL, sub = NULL,
  xlab = NULL, ylab = NULL, xaxis = TRUE, yaxis = TRUE, box = TRUE,
  contour = TRUE, contourLevels = 10, contourColor = rgb(0, 0, 0, 0.4),
  ...)
}
\description{
r.plot.heatmap
}
