% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_heatmap.R
\name{r.plot.heatmap}
\alias{r.plot.heatmap}
\title{r.plot.heatmap}
\usage{
r.plot.heatmap(y, x = NULL, xlim = c(0, 1), ylim = c(0, 1),
  heatmap = NULL, main = NULL, sub = NULL, xlab = NULL, ylab = NULL,
  xaxis = T, yaxis = T, box = T, type = "l", lwd = 1, pch = 16,
  cex = 1)
}
\description{
r.plot.heatmap
}

