% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_histogram.R
\name{r.plot.histogram}
\alias{r.plot.histogram}
\title{r.plot.histogram}
\usage{
r.plot.histogram(values, freq = TRUE, breaks = "Sturges",
  include.lowest = TRUE, right = TRUE, icol = 1, col = NULL,
  alpha = 0.5, colBorder = rgb(0, 0, 0), alphaBorder = 1, main = NULL,
  sub = NULL, xlab = NULL, ylab = NULL, labels = FALSE, ...)
}
\description{
r.plot.histogram
}
\seealso{
\code{\link{hist}} \code{\link{bkde}}

Wrappers: \code{\link{r.plot.bar}} \code{\link{r.plot.histogram}} \code{\link{r.plot.distribution}}

\code{\link{r.plot.heatmap}}

\code{\link{r.plot.matrix}} \code{\link{r.plot.matrix.communities}} \code{\link{r.plot.graph.text}}

\code{\link{r.plot.treemap}} \code{\link{r.plot.radial}}

\code{\link{r.plot.data}}

\code{\link{r.binning}} \code{\link{r.segment}} \code{\link{r.segment.target}} \code{\link{r.segment.target.table}}
}

