% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_kmeans.R
\name{r.plot.kmeans.smoothshapes}
\alias{r.plot.kmeans.smoothshapes}
\title{r.plot.kmeans.smoothshapes}
\usage{
r.plot.kmeans.smoothshapes(x, km = NULL, nclusters = 2, main = NULL,
  sub = NULL, xlab = NULL, ylab = NULL)
}
\description{
r.plot.kmeans.smoothshapes
}

