% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_graph.R
\name{r.plot.matrix.communities}
\alias{r.plot.matrix.communities}
\title{r.plot.matrix.communities}
\usage{
r.plot.matrix.communities(mat, quantileCutOff = 0.6, zlim = c(min(mat),
  max(mat)), levels = 20, palette = r.color.gradient.palette(c("white",
  rgb(0.4, 0.45, 0.8)), levels = levels), main = NULL, xlab = NULL,
  ylab = NULL, sub = NULL, ...)
}
\description{
r.plot.matrix.communities
}

