% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_graph.R
\name{r.plot.matrix}
\alias{r.plot.matrix}
\title{r.plot.matrix}
\usage{
r.plot.matrix(inputMatrix, zlim = NULL, pal = NULL, main = "Matrix",
  xlab = NULL, ylab = "", ...)
}
\description{
r.plot.matrix
}

