% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_matrix.R
\name{r.plot.matrix}
\alias{r.plot.matrix}
\title{r.plot.matrix}
\usage{
r.plot.matrix(mat, zlim = c(min(mat), max(mat)), palette = NULL,
  centerZero = TRUE, colorPos = rgb(70/255, 130/255, 180/255),
  colorNeg = rgb(165/255, 42/255, 42/255), main = NULL, xlab = NULL,
  ylab = NULL, sub = NULL, ...)
}
\description{
r.plot.matrix
}

