% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_matrix.R
\name{r.plot.matrix}
\alias{r.plot.matrix}
\title{r.plot.matrix}
\usage{
r.plot.matrix(mat, zlim = c(min(mat), max(mat)), palette = NULL,
  centerZero = TRUE, colorPos = rgb(70/255, 130/255, 180/255),
  colorNeg = rgb(165/255, 42/255, 42/255), main = NULL, xlab = NULL,
  ylab = NULL, sub = NULL, margin = 0.5, marginX = margin,
  marginY = margin, ticksX = NULL, ticksY = NULL, numTicks = 12,
  numTicksX = numTicks, numTicksY = numTicks, ...)
}
