% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.new}
\alias{r.plot.new}
\title{r.plot.new}
\usage{
r.plot.new(x = NULL, y = NULL, xlim = c(0, 1), ylim = c(0, 1),
  background = T, grid = T, backgroundCol = param.color.background,
  foregroundCol = param.color.foreground, axisCol = param.color.axis,
  boxCol = param.color.box, main = NULL, sub = NULL, xlab = NULL,
  ylab = NULL, xaxis = T, yaxis = T, box = T, log = "", asp = NA,
  thirdAxis = FALSE, restore.par = TRUE, ...)
}
\arguments{
\item{x}{Array of data for x axis. It will not be plotted, however it is used to calculate limits for x axis according to minimum and maximum values of data x.}

\item{y}{Array of data for y axis. It will not be plotted, however it is used to calculate limits for y axis according to minimum and maximum values of data y.}
}
\description{
Creates a new plot with no data at all.
}
\details{
﻿
}
\seealso{
Main: \code{\link{r.plot}} \code{\link{r.plot.add}} \code{\link{r.plot.new}}

3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}

Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

