% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_radial.R
\name{r.plot.radial}
\alias{r.plot.radial}
\title{r.plot.radial}
\usage{
r.plot.radial(data, grid.min = NULL, grid.mid = NULL, grid.max = NULL,
  main = NULL, sub = NULL, xlab = colnames(data)[-1], ylab = NULL,
  label.center.y = FALSE, grid.label.size = 3,
  gridline.label.offset = NULL, rounding = NULL,
  label.gridline.min = TRUE, label.gridline.mid = NULL,
  label.gridline.max = TRUE, label.color = param.color.labelaxis,
  axis.label.offset = 1.15, axis.label.size = 3, group.line.width = 0.1,
  group.point.size = 1, background.circle.color = param.color.background,
  legend = if (nrow(data) > 1) TRUE else FALSE, legend.title = NULL,
  legend.text.size = grid.label.size)
}
\description{
Radial plot done by Paul Williamson (http://pcwww.liv.ac.uk/~william/)
}

