% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot_radial.R
\name{r.plot.radial}
\alias{r.plot.radial}
\title{Radial plot done by Paul Williamson (http://pcwww.liv.ac.uk/~william/)}
\usage{
r.plot.radial(plot.data, axis.labels = colnames(plot.data)[-1],
  grid.min = -0.5, grid.mid = 0, grid.max = 0.5, centre.y = grid.min -
  ((1/9) * (grid.max - grid.min)), plot.extent.x.sf = 1.2,
  plot.extent.y.sf = 1.2, x.centre.range = 0.02 * (grid.max - centre.y),
  label.centre.y = FALSE, grid.line.width = 0.5,
  gridline.min.linetype = "longdash", gridline.mid.linetype = "longdash",
  gridline.max.linetype = "longdash", gridline.min.colour = "grey",
  gridline.mid.colour = "blue", gridline.max.colour = "grey",
  grid.label.size = 4, gridline.label.offset = -0.02 * (grid.max -
  centre.y), label.gridline.min = TRUE, axis.label.offset = 1.15,
  axis.label.size = 3, axis.line.colour = "grey", group.line.width = 1,
  group.point.size = 4, background.circle.colour = "yellow",
  background.circle.transparency = 0.2, plot.legend = if (nrow(plot.data) >
  1) TRUE else FALSE, legend.title = "Cluster",
  legend.text.size = grid.label.size)
}
\description{
r.plot.radial
}

