% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot}
\alias{r.plot}
\title{r.plot}
\usage{
r.plot(x = NULL, y = NULL, xlim = c(0, 1), ylim = c(0, 1),
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL, xaxis = T,
  yaxis = T, box = T, type = "p", lwd = 1, pch = 16, cex = 1,
  icol = NULL, col = NULL, alpha = NULL, log = "", asp = NA,
  thirdAxis = FALSE, ...)
}
\description{
r.plot
}
\seealso{
\code{\link{r.plot.new}} \code{\link{r.plot}} \code{\link{r.plot.add}}
\cr 3rd axis: \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
\cr Colors: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}} \code{\link{r.color}} \code{\link{r.color.gradient}}
\cr Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

