% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot}
\alias{r.plot}
\title{r.plot}
\usage{
r.plot(x = NULL, y = NULL, xlim = c(0, 1), ylim = c(0, 1),
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL, xaxis = T,
  yaxis = T, box = T, type = "p", lwd = 1, pch = 16, cex = 1,
  icol = NULL, col = NULL, alpha = NULL, log = "", asp = NA,
  thirdAxis = FALSE, ...)
}
\arguments{
\item{x}{Array of data for x axis.}

\item{y}{Array of data for y axis.}
}
\description{
Creates a new plot and plots some data.
}
\details{
﻿
}
\seealso{
Main: \code{\link{r.plot}} \code{\link{r.plot.add}} \code{\link{r.plot.new}}

3rd axis: \code{\link{r.plot.coord}} \code{\link{r.plot.coord.axis}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}

Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}

