% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplot.R
\name{r.plot.reset}
\alias{r.plot.reset}
\title{r.plot.reset}
\usage{
r.plot.reset()
}
\description{
r.plot.reset
}
\seealso{
Tools: \code{\link{r.plot.window}} \code{\link{r.plot.reset}} \code{\link{r.plot.close}} \code{\link{r.divideplot}}
}

