% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot_model.R
\name{r.plot.table}
\alias{r.plot.table}
\title{r.plot.table}
\usage{
r.plot.table(t, big.mark = ".", decimal.mark = ",", scientific = FALSE,
  paging = FALSE, searching = FALSE, bInfo = FALSE, bSortable = FALSE)
}

