\name{r.plot.window}
\alias{r.plot.window}
\title{r.plot.window}
\usage{
  r.plot.window(width = 900, height = 900, ...)
}
\description{
  r.plot.window
}
\seealso{
  \code{\link{r.plot.new}} \code{\link{r.plot}}
  \code{\link{r.plot.add}} \cr 2nd axis:
  \code{\link{r.plot.setmargins}}
  \code{\link{r.plot.newaxis}} \code{\link{r.plot.addaxis}}
  \cr Colors: \code{\link{r.palette}} \code{\link{r.color}}
  \cr Tools: \code{\link{r.plot.window}}
  \code{\link{r.plot.reset}} \code{\link{r.plot.close}}
}

