\name{r.plot2D.data}
\alias{r.plot2D.data}
\title{r.plot2D.data}
\usage{
  r.plot2D.data(x, clustReal = NULL, clustModel = NULL,
    comp1 = 1, comp2 = 2, xlim = c(0, 1), ylim = c(0, 1),
    main = "", sub = "", xlab = "", ylab = "", xaxis = T,
    yaxis = T, box = T)
}
\description{
  r.plot2D.data
}

