% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rplot_dimred.R
\name{r.plot2D.nn}
\alias{r.plot2D.nn}
\title{r.plot2D.nn}
\usage{
r.plot2D.nn(x, plotNN = TRUE, stepmax = 10^8, threshold = 0.1,
  clustReal = NULL, clustModel = NULL, xlim = c(0, 1), ylim = c(0, 1),
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL, xaxis = T,
  yaxis = T, box = T)
}
\description{
r.plot2D.nn
}

