% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{r.setPalette}
\alias{r.setPalette}
\title{r.setPalette}
\usage{
r.setPalette(palette)
}
\seealso{
\code{\link{r.palette.set}}

Palette: \code{\link{r.palette.set}} \code{\link{r.palette.restore}} \code{\link{r.palette.get}}

Colors: \code{\link{r.color}} \code{\link{r.color.gradient}} \code{\link{r.color.gradient.palette}}

Defaults: \code{\link{r.setAlpha}} \code{\link{r.setPalette}}
}

