% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_modifications.R
\name{read_modifications}
\alias{read_modifications}
\title{read_modifications}
\usage{
read_modifications(xmlurl, modifications)
}
\arguments{
\item{xmlurl}{modifications.xml file path}

\item{modifications}{modifications}
}
\value{
modifications
}
\description{
read modifications.xml file
}
\author{
Yiming Zhao
}
